Mostly pirate profile view related
Moved all of the pirate and captain view scss to captain_pirate.css.scss instead of user.css.scss
	renamed:    app/assets/stylesheets/user.css.scss -> app/assets/stylesheets/captain_pirate.css.scss

Added a bunch of comments... We need to convert all of the px to em or %'s
	modified:   app/assets/stylesheets/index.css

Moved all of the navbar related css to the same place. Made it fixed instead of absolute
	modified:   app/assets/stylesheets/nav_bar.css

Removed anything that was not general and put them in the captain_pirate.css.scss file.
	modified:   app/assets/stylesheets/user.css.scss

Added @tasks and @treasures to the show method for the view.
	modified:   app/controllers/pirates_controller.rb

Added a container to separate the navbar from all of the pages.
	modified:   app/views/layouts/application.html.haml

Made the pirate profile view responsive.
	modified:   app/views/pirates/show.html.haml